requirements.txt
streamlit==1.28.1
langchain==0.0.335
langchain-openai==0.0.2
openai==1.3.7
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.3
numpy==1.25.2
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
scikit-learn==1.3.2
plotly==5.17.0
streamlit-aggrid==0.3.4
selenium==4.15.2
webdriver-manager==4.0.1
pydantic==2.5.0
json5==0.9.14